move next-question button up a little in in-video surveys, and add titles to confirmation dialogs
The "next question" button in in-video surveys used to bump right up
against the bottom of the survey - I've moved it a little lower so that
it doesn't - tested with changing the length of the survey to confirm
it still looked good.  Also: the "survey completed" and "csv exam
completed" dialogs had blank title strings before, which was a little
odd-looking.  Added titles.